refactor: unify naming of new plugins
